+++ M11_PairProgramming_Floristeria_N2/src/com/nivel2/controller/removeProdMenu/RemoveTreeController.java
@@ -14,25 +14,28 @@ import com.nivel2.view.ShowInfoWindow;
 import com.nivel2.view.utils.Session;
 
 public class RemoveTreeController extends Controller{
+
 	FloristRepository floristRepository;
 	
 	public RemoveTreeController(Session session) {
 		super(session);
 		this.floristRepository = FloristRepository.instance();
 	}
-	
+
 	public void control() {
 		int activeFloristId = ActiveFlorist.instance().getId();
-		ShowInfoWindow.showInfo(floristRepository.getProducts(activeFloristId)
-				.stream().filter(s -> s instanceof Tree).map(l -> l.toString()).collect(Collectors.toList()));
-		
-		
-		Product product = floristRepository.getProductById(activeFloristId, this.getProdId());
-		if(product instanceof Tree) {
-			this.floristRepository.remove(product, activeFloristId);
-			ShowInfoWindow.showInfo("ARBOL ELIMINADO");
+		List<String> treeStock = floristRepository.getProducts(activeFloristId).stream()
+				.filter(s -> s instanceof Tree).map(l -> l.toString()).collect(Collectors.toList());
+		ShowInfoWindow.showInfo(treeStock);
+		if (treeStock.isEmpty()) {
+			ShowInfoWindow.showInfo("NO HAY PRODUCTOS PARA RETIRAR DE ESTE STOCK");
+		} else {
+			Product product = floristRepository.getProductById(activeFloristId, this.getProdId());
+			if (product instanceof Tree) {
+				this.floristRepository.remove(product, activeFloristId);
+				ShowInfoWindow.showInfo("ARBOL ELIMINADO");
+			}
 		}
-		
 	}
 	private int getProdId() {
 		return ReadInfoWindow.readIdProduct(MessageView.CHOOSE_PRODUCT, this.getIdList());
@@ -43,3 +46,4 @@ public class RemoveTreeController extends Controller{
 		return this.floristRepository.getProducts(activeFloristId).stream().filter(s-> s instanceof Tree).map(s -> s.getId()).collect(Collectors.toList());
 	}
 }
+
